SSM-3564: fix integer overflow when converting seconds to milliseconds
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@29890 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/gateway/assertions/SshAssertion/src/main/java/com/l7tech/external/assertions/ssh/console/SshRouteAssertionPropertiesPanel.java
@@ -147,8 +147,8 @@ public class SshRouteAssertionPropertiesPanel extends AssertionPropertiesOkCance
         validators.addRule(validators.constrainTextFieldToBeNonEmpty(getResourceString("hostNameLabel"), hostField, null));
         validators.addRule(validators.constrainTextFieldToBeNonEmpty(getResourceString("fileNameLabel"), fileNameTextField, null));
         validators.addRule(validators.constrainTextFieldToBeNonEmpty(getResourceString("usernameLabel"), usernameField, null));
-        validators.addRule(validators.constrainTextFieldToNumberRange(getResourceString("sshTimeoutLabel"), connectTimeoutTextField, 1, Integer.MAX_VALUE));
-        validators.addRule(validators.constrainTextFieldToNumberRange(getResourceString("sshTimeoutLabel"), readTimeoutTextField, 1, Integer.MAX_VALUE));
+        validators.addRule(validators.constrainTextFieldToNumberRange(getResourceString("sshTimeoutLabel"), connectTimeoutTextField, 1, Short.MAX_VALUE));
+        validators.addRule(validators.constrainTextFieldToNumberRange(getResourceString("sshTimeoutLabel"), readTimeoutTextField, 1, Short.MAX_VALUE));
 
         validators.addRule(new InputValidator.ComponentValidationRule(portNumberTextField) {
             @Override
